pyproject.toml
[project]
name = "mang2vec"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "cairosvg>=2.8.2",
    "hydra-core>=1.3.2",
    "numpy>=2.2.6",
    "opencv-python>=4.12.0.88",
    "torch>=2.8.0",
]
